data/BACKEND/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
dj-rest-auth==7.0.0
Django==5.1.2
django-allauth==65.2.0
djangorestframework==3.14.0
gunicorn
idna==3.10
packaging==24.1
pillow==11.0.0
psycopg2-binary==2.9.10
Pygments==2.18.0
pytz==2024.2
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
django-prometheus
django-jazzmin
stripe
asgiref
twisted[tls,http2]
colorlog
django-environ
